requirements.txt
torch==2.2.2
diffusers==0.3.0
python-dotenv==1.0.1
